+++ src/main/java/com/bsuiramt/servetogetherbackend/dto/response/GroupInformationDTO.java
@@ -1,0 +1,6 @@
+package com.bsuiramt.servetogetherbackend.dto.response;
+
+import java.util.List;
+
+public record GroupInformationDTO(List<String> members) {
+}
