+++ src/main/java/com/bsuiramt/servetogetherbackend/exception/AnnouncementIsNotAvailable.java
@@ -1,0 +1,4 @@
+package com.bsuiramt.servetogetherbackend.exception;
+
+public class AnnouncementIsNotAvailable {
+}
